test: Use an environment variable for the log file

+++ src/fourdvar/logging.py
@@ -20,8 +20,20 @@ import os
 from fourdvar.params.root_path_defn import root_path, store_path
 
 
-def setup_logging(verbose: bool = False):
-    """Setup logging for the project."""
+def setup_logging(verbose: bool = False, reset_logfile: bool = True):
+    """
+    Setup logging for the project.
+
+    The `OM_LOGGING_FILE` environment variable can be set to specify the name of the log file.
+    This file will be created in the `store_path` directory.
+
+    Parameters
+    ----------
+    verbose
+        If True, set file logging level to DEBUG, otherwise set to INFO.
+    reset_logfile
+        If True, delete the log file if it already exists.
+    """
     to_screen_level = logging.INFO
     to_file_level = logging.DEBUG if verbose else logging.INFO
 
@@ -30,32 +42,29 @@ def setup_logging(verbose: bool = False):
     # to_file_format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
     to_file_format = "%(name)s - %(levelname)s - %(message)s"
 
-    logfile_name = "full_report.log"
-
-    reset_logfile = True
-
     project_name = os.path.split(root_path)[1]
-
-    logfile = os.path.join(store_path, logfile_name)
-
-    if reset_logfile is True and os.path.isfile(logfile):
-        os.remove(logfile)
 
     base_logger = logging.getLogger(project_name)
     base_logger.setLevel(logging.DEBUG)
 
     to_screen_handle = logging.StreamHandler()
-    to_file_handle = logging.FileHandler(logfile)
     to_screen_handle.setLevel(to_screen_level)
-    to_file_handle.setLevel(to_file_level)
-
     to_screen_formatter = logging.Formatter(to_screen_format)
-    to_file_formatter = logging.Formatter(to_file_format)
-
     to_screen_handle.setFormatter(to_screen_formatter)
-    to_file_handle.setFormatter(to_file_formatter)
     base_logger.addHandler(to_screen_handle)
-    base_logger.addHandler(to_file_handle)
+
+    if os.environ.get("OM_LOGGING_FILE"):
+        logfile_name = os.environ.get("OM_LOGGING_FILE")
+        logfile = os.path.join(store_path, logfile_name)
+
+        if reset_logfile is True and os.path.isfile(logfile):
+            os.remove(logfile)
+
+        to_file_handle = logging.FileHandler(logfile)
+        to_file_handle.setLevel(to_file_level)
+        to_file_formatter = logging.Formatter(to_file_format)
+        to_file_handle.setFormatter(to_file_formatter)
+        base_logger.addHandler(to_file_handle)
 
     base_logger.debug("Logging setup finished.")
